requirements.txt
absl-py==0.15.0
aspose-barcode-for-python-via-net==23.7
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.2.4
certifi==2023.7.22
charset-normalizer==2.0.12
clang==5.0
cycler==0.11.0
dataclasses==0.8
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-images-download==2.8.0
google-pasta==0.2.0
grpcio==1.48.2
h5py==3.1.0
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.7
matplotlib==3.3.4
numpy==1.19.5
oauthlib==3.2.2
opencv-python==4.8.0.76
opt-einsum==3.3.0
pandas==1.1.5
Pillow==8.4.0
pkg_resources==0.0.0
protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydot==1.4.2
pyparsing==3.1.1
python-barcode==0.15.1
python-dateutil==2.8.2
pytube==12.1.3
pytube3==9.6.4
pytz==2023.3.post1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
selenium==3.141.0
simple-image-download==0.2
six==1.15.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-estimator==2.6.0
termcolor==1.1.0
tqdm==4.64.1
typing-extensions==3.7.4.3
urllib3==1.26.16
Werkzeug==2.0.3
wrapt==1.12.1
zipp==3.6.0
